Add outcome group header to the Find Outcome modal
When an individual outcome group is clicked on the left, render
a header on the right hand side of the modal. Within the header,
include the # of sub-outcomes and nonfunctional
‘Add All Outcomes’ button

closes OUT-4150
flag=improved_outcomes_management

Test plan:
- Make sure you have Account and Course with outcome groups
- Go to Account > Settings > Feature Options
- Enable Improved Outcomes Management FF
- Go to Account > Outcomes
- click on Find button; a fullscreen modal should open
and display billboard with PRO TIP in the right panel
- select an individual outcome group via tree browser on the left
- the right panel should display outcome group header with group
title and number of sub-outcomes.
- there should be a nonfunctional buttton "Add All Outcomes"
- close the modal and reopen it; you should see the billboard
with PRO TIP in the right panel
- Go to Courses > select a course > Outcomes
- repeat the same tests for the Course as for the Account
- confirm that the visualization and responsiveness of outcome
group header is similar under Chrome, Firefox and Safari

Change-Id: Ie2c51ca5ddc0cff3fa289fb3b260344b45e9ca60
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/257361
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Pat Renner <prenner@instructure.com>
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
QA-Review: Michael Brewer-Davis <mbd@instructure.com>
Product-Review: Jody Sailor