Always use primary DB when doing LTI assignment fixups
Closes INTEROP-6442
flag=none

With LTI 1.1 -> 1.3 upgrade path, we do a "just in time"
fixup of LTI 1.1 assignments by adding an Lti::ResourceLink
and the default LineItem when the assignment is shown.

The assignment show action wraps this code in a read-only
transaction, which causes issues when we attempt to create
the needed records.

This change ensures Canvas uses the primary DB when creating
the required records.

Test Plan:
- Set up a Canvas secondary user
- Create an LTI 1.3 assignment
- In the rails console, delete the assignment's resource link and
  line items:
  ```
  <assignment>.primary_resource_link.delete
  <assignment>.line_items.delete_all
  ```
- Navigate to the assignment and verify the tool launch
  occurs properly
- Verify a new line item and resource link were created for
  the assignment

Change-Id: I8acaa094bd43ea5fb4fc44f1011b7673665b635e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258373
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Weston Dransfield <wdransfield@instructure.com>