allow sending messages for account context x shard
the issue was that it was checking for account assocations on one shard
when it is living on another. This simplifies to check on all the users
shards but scoped to the specific root_account from the where clause.

added spec covers the case

test plan
 - have sharding setup
 - have user from other shard as a sub_account admin
 - go to /users/ at the sub_account level as the admin
 - attempt to message the user
 - it should not render invalid context error

fixes VICE-1183
flag=none

Change-Id: I02f1a29fd6882f8177eb7364a335c8da20b2fe22
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260414
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Davis Hyer <dhyer@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>