Update <Select> in Discussion sections editor
Closes FOO-1578
flag = none

When adding or editing a discussion, there is a multiple select
box for controlling which sections the discussion should be
posted to. This had to be converted from the deprecated InstUI
<Select> to the <CanvasMultiSelect> component.

Also had to add the ability to pass along a disabled prop to
CanvasMultiSelect.

Also had to do a lot of fiddling with I18n to reduce the bundle
size, sigh.

Test plan:
* Have a discussion in a course with multiple sections
* Edit it (/courses/:id/discussion_topics/:id/edit)
* Underneath the RCE for entering the discussion text is a
  "Post to" selection box
* That box should work as it did before

Change-Id: I2b0187d3cf101db71934445d9c76581956fb47ac
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/261798
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>