Use equation image src attribute as the source of truth
refs MAT-26
flag=new_math_equation_handling

we were pulling the equation LaTex from the image's
data-equation-content attribute but in the quiz from the ticket
its value was not the equation. I'm not sure how this happened
(could the quiz author really have edited the HTMl to change
all the quiz's equations?) but since the src UTL is the equation's
source of truth, let's just always use that.

There are other places in the code that rely on the date-equation-content
attribute to generate mathml, but it's only exercised if the
new math feature is off. I'd rather not remove it from everywhere on
a ticket that will be hotfixed.

test plan:
test plan:
  - pull stable/2021-04-14 (or maybe you can just fetch this commit
    from gerrit)
  - script/nuke_node.sh
  - get a cup of coffee
  - with the new_math_equation_handling flag on
  - in the rce use the equation editor to create an equation
  - edit the HTML to put something else in the data-equation-content
    and alt attributes, or remove all but the src attribute if you
    want.
  - save
  > expect the correct MathJax typset equation in the result.
  - have an equation with no LaTex in the src attribute and save
  > expect the broken-image image and no catastrophic failures

Change-Id: I3ac4fb5dfcbf9557dfc8551d802e19dcc965c7f7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/262566
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>