Add create new group component at bottom of move modal
closes OUT-4213
flag=improved_outcomes_management

test plan:
- Prereq: FF turned on and Account/Course with groups and
outcomes
- Navigate to Account with Outcomes
- Click on a root outcome group from the tree browser on the
left-hand side
- Click on Kebab Menu next to Outcome Group title
- Click on the Move option; this will open the Move Modal
- You should see the “+ Add new group” button displayed
below the Group Tree Browser
- Click on the “+ Add new group” button, a form will display
with a text input, a checkmark button and a X button. Note
that the checkmark button should remain disabled until a
group name is provided in the text area.
- Provide a new group name in the text area.  Click the X button.
- This should close the form and revert back to displaying
 “+ Add new group” button
- Click on the “+ Add new group” button again and the text area
Should be cleared of the text you entered previously.
- Provide a new group name that you want to create. Note that
the Checkmark button should enable.
- Click on the checkmark button to save the group.  This should
revert back to showing “+ Add new group” button and once
saved, a flash alert will display showing that the group was added.
NOTE: The flash alert is only to verify that the group was saved.
This is not a part of the designs and will be removed later.
- Refresh the index page and you should see the new group was
added to the root.
- Repeat the same steps in a Course with outcomes

Change-Id: Ie82ff9b16615c496475dc44b90a6f61ad20ea424
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/261893
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
QA-Review: Martin Yosifov <martin.yosifov@instructure.com>
Product-Review: Jody Sailor