MSFT Sync: retry 404s in execute_diff stage
When we get a 404 in the execute_diff stage, go back to the
generate_diff stage.

Because we're now essentially retrying generate_diff & execute_diff as
one big step, it should just be one step, but we're going to add
separate progress-capture retry soon so we'd have to split it up again
then. And, testing them separately still in the meantime is nice. So it
was easier just to add "Retry with step".

That said, I've left this a separate commit in case we want to remove
support with "Retry with step" later.

closes INTEROP-6638
flag=microsoft_group_enrollments_syncing

Test plan:
- At the beginning of step_execute_diff, add:
  if group.reload.job_state&.dig(:retries_by_step)&.dig('step_generate_diff').to_i < 3
    raise MicrosoftSync::Errors::HTTPNotFound.new(
      service: 'foo', response: Struct.new(:body, :code).new('foo', 404), tenant: 'foo'
    )
  end
- Run job synchronously (g.syncer_job.run_synchronously) and watch logs
  to see that it retries 3 times, doing back to generate_diff from
  execute_diff, sleeping the specified time for retry, before succeeding.
- Run job asynchronously (g.syncer_job.run_later) and check the same.

Change-Id: I366c959ec8198801788b8b55b58847fd7329c439
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263048
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>