Fix duplicate requests on K-5 dashboard
When the dashboard and k5_dashboard bundles are loaded, the dashboard
cards are rendered twice-- once with their state from sessionStorage,
then again with the actual authoritative response from the server. A
lot of related requests (homeroom announcements, subject course
announcements, and LTIs) are fired in response to those renders, which
can lead to spurious errors when the first of the two sets of requests
are canceled. Additionally, the requests based on the cached
sessionStorage card values can actually be incorrect (e.g. requesting
the announcements and LTIs from a course that a user was just
unenrolled from).

So this change updates those dependent requests to only fire after
we've gotten the authoritative server response with the dashboard card
info. It adds a `cardsSettled` param to the dashboard cards render
function to accomplish this.

fixes LS-2172
flag = canvas_for_elementary

Test plan:
  - Log in to the K-5 dashboard as a student
  - Expect the cards and various tabs to work as they did before
  - In another session, unenroll the student from one of their courses
  - Make sure you have Chrome dev tools open (or equivalent) and ready
    to capture network requests
  - As the student again, reload the dashboard
  - Expect to not get any 401 unauthorized errors in the console
  - Expect to see no duplicate requests for announcements and LTIs for
    the same course ID (e.g. /api/v1/courses/<id>/external_tools and
    /api/v1/announcements?context_codes=course_<id>)

Change-Id: Ibcd245eb1bedb9288720d49930b43eed58f408c1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263534
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Jeff Largent <jeff.largent@instructure.com>