add telemetry support for zsh
Zsh is the default shell on the new macOS and does not
implement function export features. In order to have both
bash and zsh to work we changed the snippet for opting-in
telemetry using a variable to store the function.

test-plan:
- local docker smoke passes
- setup/update scripts work in bash
- setup/update scripts work if executed on zsh
- test enable/disable telemetry from .bashrc and .zshrc

Change-Id: I4e034f486dde078bba517172114e9b3e01f1817f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/264024
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
QA-Review: Andrea Cirulli <andrea.cirulli@instructure.com>
Product-Review: Andrea Cirulli <andrea.cirulli@instructure.com>