fix flakey gradebook spec
refs DE-665

When entering in a grade with a thousands separator, the actual behaviour is that it switches first to an entry with no comma, then the comma entry after that. Thus, this test is looking for an intermediate element, so fix it by switching to looking for the final rendered element instead.

Change-Id: I153b14ca4bae493448a0f1c979a27e86db8b16fe
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/264397
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Andrea Cirulli <andrea.cirulli@instructure.com>
QA-Review: Aaron Ogata <aogata@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>