fail build if retry doesn’t run any tests
refs DE-686

Fix an issue where a test retry run with no tests could result in an erroneously passing build. The issue was that the rspec formatter was failing and rspec would not output the only-failures state log. This caused the next only-failures call to not run any tests, but claim to succeed even though 0 tests were run. This issue was fixed by failing rspec upon 0 tests run.

Also, use [skip-stages=Flakey] because the test exceeds the timeout 1/10 runs, and is out-of-scope for this ticket to fix.

Test Plan
1. Retries fail when they run zero tests
2. Retries work correctly when they have real failures + formatter failures (they should run 0 tests and fail)
3. Retries work when multiple threads in a single stage have failures

Change-Id: Ibc790f4a102fb86c1ffdab4a49eaf3b5f1452631
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/264788
QA-Review: Aaron Ogata <aogata@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Andrea Cirulli <andrea.cirulli@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>