Tweak global nav strings for k5
Creates a helper in the application_controller called `k5_user?` which
returns true iff the user is associated with any k5 accounts. Also sets
a JS_ENV variable called :K5_USER with the value of ks_user?. Renames
other JS_ENV variables to :K5_HOMEROOM_COURSE and :K5_SUBJECT_COURSE for
clarity. Also removes references to 'k5_mode', replacing with calls to
k5_user? or calls on the Course model for clarity.

use_k5? is cached since it is frequently referenced. The cache key is
cleared for each user in the account when k5 mode is toggled by an admin
or when enrollments are adjusted. It is also set to expire after 1 hour,
which might not be necessary. Cache is not invalidated when the k5 flag
is toggled since the flag is being removed.

The user dashboard and global nav states are dependent on use_k5? Shows
`Subjects` instead of `Courses` and `Homeroom` instead of `Dashboard` in
the global nav when k5 is on. Also filters homeroom courses out of
course list when k5 is on.

closes LS-1827
flag=canvas_for_elementary

[fsc-max-nodes=20]
[fsc-timeout=40]

Test plan:
 - Login as a teacher enrolled in at least one course in a k5 account
 - Expect to see k5 dashboard
 - Expect to see `Subjects` and `Homeroom` (with home icon) in the
   global nav (and in the mobile global nav)
 - Toggle the courses/subjects tray in global nav and expect to see
   only references to Subjects
 - Open /courses/ and expect the page heading to be `Subjects`
 - Browse through the course and expect it to look normal (w/ Balsamiq)
 - Switch to a student and expect the above to be true, plus the courses
   list should not include homeroom courses
 - Disable k5 for the account and return to teacher or student
 - Expect to immediately see classic canvas nav, dashboard, and courses
   again
 - Test caching by placing a log statement (i.e.
   `p "K5 Cache: uncached"`) in ApplicationController#uncached_use_k5?
 - Run `redis-cli FLUSHDB`
 - Navigate around canvas; expect to see the log statement only once in
   the logs until toggling k5 mode or adding the first k5 course to a
   user

Change-Id: Iec970c939b976d10eee13d40b7bf68f601d8178e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/265239
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>