Add an external plugin  to the demo app
closes MAT-196
flag=none

The canvas-independent rce needs an api to let clients
add plugins and place them where they want in the toolbar
and menus.  This adds this function to the RCE and the demo
app, though I'm not convinced it's the correct API. It does
show how it can be done though.

test plan:
  - yarn demo:dev
  - check the "Include test plugin" checkbox
  - pick a toolbar
  - pick a menu
  - click Update
  > expect a wrench button in the selected toolbar and menu
  - click the wrench
  > expect an alert

You can select one of the existing toolbars or menus,
or the (new!) option and it will create a new
toolbar or menu.

Change-Id: I4dde32e136db9ee922e7f183aae165c6c40c6461
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/265293
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Guilherme Baron <gbaron@instructure.com>
QA-Review: Guilherme Baron <gbaron@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>