Fix html editor for assignment enhanced student view
Closes MAT-151
flag=none

The primary issue here was that the old code was
attempting to add a "focus" event handler to the
RCE iframe with "querySelector('[id^="random_editor"]')".

Because this query did not select an element type, the
TextArea with "random_editor_id_<id>" was selected
and had the "focus" event handler attached to it rather
than the intended iframe.

I opted to just delete all the iframe focus managment.
Doing a before / after test shows that focus managment
with keyboard or mouse is working the same. This makes
sense since the iframe was not getting the event listner
in the first place.

Test Plan:
- Enable assignments enhanced student view for a course
- Create an assignment in that course of type online
  text entry and file upload
- view the assignment as a student
- Select "Text" for submission type and enter some
  text into the RCE if it's not already selected
- Verify the RCE receives focus
- Switch to both the raw and pretty HTML editors and
  verify they allow modyfing the HTML content

Change-Id: Ifa8c6e65226c5be04a04011ed1868a6be1506248
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/266563
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Weston Dransfield <wdransfield@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>