Subject Home empty state
A component called EmptyHome was added, it will be shown when the
course does not have a front page set, if the user has teacher role,
a manage home button will also be rendered to allow the user add
or select a front page. This change also renames the “Front Page”
indicator in the pages list to “Subject home”, as well as the options
in the three-dot menu, which now make reference to "Subject Home"
instead of "Front Page"

fixes LS-2296
flag=none

Test plan:

- As student go to a course that does not have front page set,
and expect to the panda be presented with the following text
bellow:
- “‘This is where you’ll land when your home is complete.”
- As teacher go to the same course or any other without a front page
set, and expect:
- The same panda and text be presented as in the student
scenario
- A button “Manage Home” to be presented, click that button and
expect:
- A redirection to the course pages list to select a front page if
the	course has pages already.
- A redirection to a new page creation if the course does not
have any page

Change-Id: If70510e849ad5c3bb1c338fe93c005ef79816bb7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/267056
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>