allow creating courses with nightly_sync
when this feature was made the update was copied to the create.
The create didn't have a @course object and it is not yet saved either.
The first use of @course is to assign the attribute. The second was for
a permission check that will fail since it is not saved, and if the user
has permissions to create the course they have permissions to set the
setting.

test plan
 - specs should pass

fixes VICE-1634
flag=none

Change-Id: I44bddf759665edf3b52ff22df72c6bc5d5480b93
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/267622
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>