Make media comments work with group assignments
When submitting a media comment from the submission details page, pass
an empty string for the value of "comment" to let the server know that
we are in fact submitting a comment. This fixes a bug with media
comments for group assignments where the comments were not being
properly saved for the whole group (and which didn't occur when
commenting from SpeedGrader because it already passed an empty string).

fixes EVAL-1729
flag=none

Test plan:
- Have a Canvas installation with Notorious set up
- Create a group assignment
- As the teacher, check that you can leave media comments both from
  SpeedGrader and the submission details page, and they appear properly
  (and are replicated to other students in the group) instead of just
  silently failing
- Smoke-test plain old text comments and media comments for non-group
  assignments on both pages

Change-Id: Ie36243b4f69c98177a6322cf9ebfb129b750f944
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/268306
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>