apply apollo-federation base classes to data graph
refs INTEROP-6900

these changes are based on the guide at
https://github.com/Gusto/apollo-federation-ruby#getting-started
even without these changes, the GraphQL data graph is already able to be
registered in a federated apollo server.  the value the gem provides is
a declarative interface for extending our schema with "entities"
(https://www.apollographql.com/docs/federation/entities/) which allow
subgraphs in a federation to reference and extend types defined in other
services' subgraphs.

so, if and when the api gateway is operational and other services start
publishing data graphs that need to reference objects in canvas' graph,
this work will be the foundataion to enable that.

at this point there should be no functional difference to canvas.  the
GraphQL API should be working the same as always.

Change-Id: I5e0de265ff85ecff53e1066ca0cb23f9f349a90f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/268714
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Michael Ziwisky <mziwisky@instructure.com>
Product-Review: Michael Ziwisky <mziwisky@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>